Move quotes-api to quote directory
