+++ src/modules/book/book.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { GoogleApiModule } from '../google-api/google-api.module';
+import { GoogleApiModule } from './api/google-api.module';
 import { TokenModule } from '../token/token.module';
 import { ValidationModule } from '../validation/validation.module';
 import { BookController } from './book.controller';
