+++ src/modules/quote/types/IQuote.ts
